LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_1-img-src-expected.txt

CONSOLE MESSAGE: Refused to load http://www1.localhost:8800/content-security-policy/support/fail.png because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
default-src should cascade to img-src directive


Harness Error (TIMEOUT), message = null

PASS Verify cascading of default-src to img-src policy
NOTRUN Should fire violation events for every failed violation
